# perf(http): increase performance of HTML rendering by ~14x