```ts
Deno.bench("html", () => {
  html`
    <div class="test">Hello, World!</div>
  `;
});
```

Before:
```
benchmark   time/iter (avg)        iter/s      (min … max)           p75      p99     p995
----------- ----------------------------- --------------------- --------------------------
html               108.0 ns     9,260,000 (104.0 ns … 125.4 ns) 109.6 ns 119.6 ns 122.0 ns
```

After:
```
benchmark   time/iter (avg)        iter/s      (min … max)           p75      p99     p995
----------- ----------------------------- --------------------- --------------------------
html                 7.7 ns   130,000,000 (  7.2 ns …  22.3 ns)   7.7 ns   9.5 ns  10.1 ns
```